:art: パーサのロジック部分を複数のソースファイルに分離した
