+++ crates/parser/src/lib.rs
@@ -9,86 +9,8 @@
 //! term ::= INT_LIT | STRING_LIT ;
 //! ```
 
+mod parse;
 mod result;
 
-use itertools::{put_back, PutBack};
-use mylang_ast::{Expr, Stmt};
-use mylang_token::{KeywordKind, Locatable, Pos, Token};
-
+pub use parse::parse;
 pub use result::{ParseErr, ParseResult};
-
-fn term(tokens: &mut PutBack<impl Iterator<Item = Token>>, pos: Pos) -> ParseResult<(Pos, Expr)> {
-    if let Some(tok) = tokens.next() {
-        match tok {
-            Token::I32(range, n) => Ok((range.end_ref().clone(), Expr::I32Lit(range, n))),
-
-            Token::Str(range, s) => Ok((range.end_ref().clone(), Expr::StrLit(range, s))),
-
-            _ => Err(ParseErr::TermExpected(tok.locate().end())),
-        }
-    } else {
-        Err(ParseErr::TermExpected(pos))
-    }
-}
-
-fn expr(tokens: &mut PutBack<impl Iterator<Item = Token>>, pos: Pos) -> ParseResult<(Pos, Expr)> {
-    let (pos, lhs) = term(tokens, pos)?;
-
-    match tokens.next() {
-        Some(Token::AddOp(_)) => {
-            let (pos, rhs) = expr(tokens, pos)?;
-            Ok(match rhs {
-                Expr::Add(b, c) => (pos, Expr::Add(Box::new(Expr::Add(Box::new(lhs), b)), c)),
-
-                _ => (pos, Expr::Add(Box::new(lhs), Box::new(rhs))),
-            })
-        }
-
-        Some(tok) => {
-            tokens.put_back(tok);
-            Ok((pos, lhs))
-        }
-
-        None => Ok((pos, lhs)),
-    }
-}
-
-fn stmt(tokens: &mut PutBack<impl Iterator<Item = Token>>, pos: Pos) -> ParseResult<Stmt> {
-    match tokens.next() {
-        Some(Token::Keyword(range, KeywordKind::PrintI32)) => {
-            let (_, expr) = expr(tokens, pos)?;
-            Ok(Stmt::PrintI32(range, expr))
-        }
-
-        Some(Token::Keyword(range, KeywordKind::PrintStr)) => {
-            let (_, expr) = expr(tokens, pos)?;
-            Ok(Stmt::PrintStr(range, expr))
-        }
-
-        Some(tok) => Err(ParseErr::KeywordExpected(tok.locate())),
-
-        _ => Err(ParseErr::KeywordExpected(pos.into())),
-    }
-}
-
-fn program(tokens: &mut PutBack<impl Iterator<Item = Token>>) -> Vec<ParseResult<Stmt>> {
-    let mut results = Vec::<Result<Stmt, ParseErr>>::new();
-
-    while let Some(tok) = tokens.next() {
-        match tok {
-            Token::Newline(_) => (),
-
-            _ => {
-                let pos = tok.locate().end();
-                tokens.put_back(tok);
-                results.push(stmt(tokens, pos));
-            }
-        }
-    }
-
-    results
-}
-
-pub fn parse(tokens: impl Iterator<Item = Token>) -> Vec<ParseResult<Stmt>> {
-    program(&mut put_back(tokens))
-}
